+++ client/Packages/com.beamable/Common/Runtime/Api/IBeamableRequester.cs
@@ -77,6 +77,11 @@ namespace Beamable.Common.Api
 		/// </summary>
 		public Func<string, T> parser;
 
+		/// <summary>
+		/// A function that accepts a set of headers, and has the opportunity to transform the headers before they are sent.
+		/// </summary>
+		public Func<Dictionary<string, string>, Dictionary<string, string>> headerInterceptor;
+
 		public Method Method => method;
 		public string Uri => uri;
 
@@ -90,6 +95,7 @@ namespace Beamable.Common.Api
 			useConnectivityPreCheck = clone.useConnectivityPreCheck;
 			parser = clone.parser;
 			disableScopeHeaders = clone.disableScopeHeaders;
+			headerInterceptor = clone.headerInterceptor;
 		}
 	}