+++ client/Packages/com.beamable/Editor/BeamCli/Commands/BeamServicesBuildCommandOutput.cs
@@ -1,0 +1,14 @@
+
+namespace Beamable.Editor.BeamCli.Commands
+{
+    using Beamable.Common;
+    using Beamable.Common.BeamCli;
+    
+    [System.SerializableAttribute()]
+    public class BeamServicesBuildCommandOutput
+    {
+        public string id;
+        public string message;
+        public bool isFailure;
+    }
+}